api_for_mcp/pdf-maker/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0
fpdf2>=2.7.0
pytest>=8.0.0
httpx>=0.27.0
